pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["VRaitzev <v.raitsev@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
sqlalchemy = { version = "^1.4", extras = ["asyncio"] }
pydantic = "^1.10.0"
python-dotenv = "^1.0.0"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
asyncpg = "^0.30.0"
psycopg2 = "^2.9.10"
alembic = "^1.14.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
python-multipart = "^0.0.19"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
